+++ client/src/monaco-languages.ts
@@ -9,10 +9,9 @@ import {
     DocumentSymbolProvider, CodeActionProvider, CodeLensProvider, DocumentFormattingEditProvider, DocumentRangeFormattingEditProvider,
     OnTypeFormattingEditProvider, RenameProvider,
     DocumentFilter, DocumentSelector, DocumentLinkProvider, ImplementationProvider, TypeDefinitionProvider, DocumentColorProvider,
-    FoldingRangeProvider,
+    FoldingRangeProvider, SemanticTokensLegend,
     DocumentSemanticTokensProvider, DocumentRangeSemanticTokensProvider
 } from "./services";
-import { SemanticTokensLegend } from 'vscode-languageserver-protocol/lib/protocol.sematicTokens.proposed'
 
 import { MonacoDiagnosticCollection } from './monaco-diagnostic-collection';
 import { ProtocolToMonacoConverter, MonacoToProtocolConverter } from './monaco-converter';
@@ -83,7 +82,7 @@ export class MonacoLanguages implements Languages {
                 const result = await provider.provideCompletionItems(params, token);
                 return result && this.p2m.asCompletionResult(result, defaultRange);
             },
-            resolveCompletionItem: provider.resolveCompletionItem ? async (model, position, item, token) => {
+            resolveCompletionItem: provider.resolveCompletionItem ? async (item, token) => {
                 const protocolItem = this.m2p.asCompletionItem(item);
                 const resolvedItem = await provider.resolveCompletionItem!(protocolItem, token);
                 if (resolvedItem) {